PTNews/requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
elastic-transport==8.12.0
elasticsearch==8.12.1
exceptiongroup==1.2.0
feedparser==6.0.11
fonttools==4.43.1
gender-guesser==0.4.0
graphviz==0.20.1
httplib2==0.20.2
importlib-metadata==4.6.4
install==1.3.5
jeepney==0.7.1
jellyfish==1.0.3
joblib==1.3.2
jusText==3.0.0
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
matplotlib==3.8.0
mercurial==6.1.1
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.0
oauthlib==3.2.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
Pillow==10.0.1
psutil==5.9.0
pyagrum==1.10.0
pydot==1.4.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==5.4.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
SecretStorage==3.3.1
segtok==1.5.11
selenium==4.18.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
systemd-python==234
tabulate==0.9.0
threadpoolctl==3.2.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.3
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.2.1
urwid==2.1.2
wadllib==1.3.6
webdriver-manager==4.0.1
wsproto==1.2.0
yake @ git+https://github.com/LIAAD/yake@0fa58cceb465162b6bd0cab7ec967edeb907fbcc
zipp==1.0.0
flask==3.0.3
blinker==1.6.2
httpx==0.13.3
idna==2.10
httpcore==0.9.1
h11==0.9.0